% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongoQueries.R
\name{absoluteHashtagCount}
\alias{absoluteHashtagCount}
\title{absoluteHashtagCount}
\usage{
absoluteHashtagCount(dataframe, interval)
}
\arguments{
\item{dataframe}{to transform}

\item{interval}{for the specified timeframe}
}
\value{
interval which has been used to run createTwitterTimeline()
}
\description{
Sums up the time series dataframe to get absolute counts
of a given time interval.
}
\examples{
#absoluteHashtagCount(dataframe, interval)
}

