% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regexBuilder.R
\name{buildRegexQuery}
\alias{buildRegexQuery}
\title{buildRegexQuery}
\usage{
buildRegexQuery(regex, timestamp, fields = c("entitites.hashtags.text",
  "text"), regexoption = "i", conjunction = "$or")
}
\arguments{
\item{regex}{the regular expression}

\item{timestamp}{to retrieve hits only within a given interval}

\item{fields}{to be searched by the regex. Default "text" and "entitites.hashtags.text".}

\item{regexoption}{options used for the regex. Default is "i", case insensitivity.}

\item{conjunction}{used to link the regex. Can only be "$and" or "$or". Default is "$or".}
}
\value{
query in jsonformat for mongolite
}
\description{
Builds a regular expression query for mongoDB in json format,
for the fields "entitites.hashtags.text" and "text". Can be use
in conjunction with a time interval.
}
\examples{
#buildRegexQuery(c("koln","koeln","berlin"), timestamp = getInterval(
#inputinterval = c("2016-01-04 18:37:43 CET","2016-01-05 18:37:43 CET")))

#hometimeline$find(buildRegexQuery(c("koln","koeln"),timestamp = getInterval(60*24)),
#getMongoFields(c("text","entitites.hashtags.text")))
}

