% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongoQueries.R
\name{cleanTwitterTimeline}
\alias{cleanTwitterTimeline}
\title{cleanTwitterTimeline}
\usage{
cleanTwitterTimeline(data, threshold = NULL)
}
\arguments{
\item{data}{to clean created by createTwitterTimeline()}

\item{threshold}{(optional) under which values will be removed from the dataframe}
}
\value{
dataframe
}
\description{
Cleans the dataframe returned by createTwitterTimeline. Inserts 0 for NaN and
removes hashtags which lie unter a certain threshold with their total amount
over all considered intervals.Threshold may be userspecified, if not will
default to an implemented threshold. See defaultThreshold().
}
\examples{
#cleanTwitterTimeline(data, threshold = NULL)
#cleanTwitterTimeline(data, threshold = 10)
}

