% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongoConnection.R
\name{connect2MongoDb}
\alias{connect2MongoDb}
\title{Establish a connection to mongoDB}
\usage{
connect2MongoDb(type)
}
\arguments{
\item{type}{Type to set the MongoDB credentials for: hometimeline or filterstream}
}
\value{
Connection Object
}
\description{
Connect to MongoDB using the credentials stored with \code{\link{setMongoDbCredentials}}
Creates either hometimeline or filterstream as a connection element.
}
\examples{
setMongoDbCredentials (type = "hometimeline", collection = "hometimeline",
db = "twitter", url = "mongodb://localhost")

}

