% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongoQueries.R
\name{createIntervalVector}
\alias{createIntervalVector}
\title{createIntervalVector}
\usage{
createIntervalVector(interval, intervalsteps)
}
\arguments{
\item{interval}{which will be divided into equally sized subintervals.
Interval must be of the form \%d.\%m.\%Y \%HH:\%MM:\%SS, for example: 11.09.2015 00:00:01}

\item{intervalsteps}{specifiy the amount of subintervals}
}
\value{
interval vector of unixtimestamps in ms
}
\description{
Creates a time interval vector. Interval parameters passed may be an unixtimestamp
(either in seconds or miliseconds), an interval of the twitter created_at format
or an interval of the form YYYY-MM-DD HH:MM:SS.
}
\examples{
#createIntervalVector(c("2015-12-15 00:00:01","2015-12-17 14:59:59"),3)
}

