% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongoQueries.R
\name{createTwitterTimeline}
\alias{createTwitterTimeline}
\title{createTwitterTimeline}
\usage{
createTwitterTimeline(mongo, interval)
}
\arguments{
\item{mongo}{a mongoDB connection to use}

\item{interval}{which will be used to aggregate the hashtags in the mongodb.}
}
\value{
dataframe containing all hashtags with count of the respective time interval
}
\description{
Creates a time series of hashtags found in the given time interval.
Is being used by createTwitterTimeSeries(). Parameter interval may consist
of an interval vector created by createIntervalVector(). See examples.
}
\examples{
#createTwitterTimeline(hometimeline,
#                        createIntervalVector(c("2015-12-15 00:00:01","2015-12-17 14:59:59"),3))
}

