% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongoQueries.R
\name{createTwitterTimeSeries}
\alias{createTwitterTimeSeries}
\title{createTwitterTimeSeries}
\usage{
createTwitterTimeSeries(mongo, interval, intervalsteps)
}
\arguments{
\item{mongo}{mongo instance to use}

\item{interval}{timeinterval}

\item{intervalsteps}{number of parts to divide the interval into}
}
\value{
dataframe
}
\description{
Creates a time series of hashtags found in the given time interval.
 The time series gets cleaned by cleanTwitterTimeline() and ordered
 in decreasing fashion with respect to the last column.
}
\examples{
#timeseries <-
# createTwitterTimeSeries(hometimeline,c("2015-12-15 00:00:01","2015-12-17 14:59:59"),3)
}

