% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongoQueries.R
\name{defaultThreshold}
\alias{defaultThreshold}
\title{defaultThreshold}
\usage{
defaultThreshold(data)
}
\arguments{
\item{data}{to calculate the threshold for.}
}
\value{
threshold integer
}
\description{
Calculates a threshlod value to remove hashtags from a dataframe
which have in sum a smaller count than the defined threshold.
Threshold is the next greatet number of the mean. Used by
cleanTwitterTimeline().
}
\examples{
#defaultThreshold(data)
}

