% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoLocation.R
\name{drawMap}
\alias{drawMap}
\title{Draw a Density Map}
\usage{
drawMap(data, location, zoom, fill = FALSE)
}
\arguments{
\item{data}{dataframe containing lat/lon coordinates}

\item{location}{Area to show on the map}

\item{zoom}{Zoom factor form the map}

\item{fill}{boolean value to adjust if a filling is used}
}
\value{
map
}
\description{
Draws a Density Map
}
\examples{
#drawMap(df, "Europe", 4)

}

