% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoLocation.R
\name{extractCoordinates}
\alias{extractCoordinates}
\title{Extract Coordinates}
\usage{
extractCoordinates(connection, searchterm, limit = 15000)
}
\arguments{
\item{connection}{The name of a MongoDB connection}

\item{searchterm}{A term all tweets returned contain}

\item{limit}{Maximum number of tweets to retrive from db}
}
\value{
dataframe with coordinates (lat,long)
}
\description{
Exports Point and Polygon coordinates from the geo fields of the tweet
The function currently supports coordinates of the type point and bounding box.
The coordinates can be shown on a map using \link{drawMap}
}
\examples{
#extractCoordinates(connection = "filterstream", searchterm = c("Trump"))

}

