% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regexBuilder.R
\name{fieldValuePair}
\alias{fieldValuePair}
\title{fieldValuePair}
\usage{
fieldValuePair(key, value, conjunction)
}
\arguments{
\item{key, }{fields to search. "text" or "entitites.hashtags.text"}

\item{value, }{regex and options as value. use operatorValuePair()}

\item{conjunction}{used for the regular expressions. Can only by "$and" or "$or"}
}
\value{
field value pair used by buildRegexQuery()
}
\description{
Creates field value pair, called by buildRegexQuery
}
\examples{
#fieldValuePair(key = c("entitites.hashtags.text","text"),value =
#operatorValuePair(c("$regex","$options"),c("syrien","i")),"$or")

}

