% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoLocation.R
\name{geocodeDataFrame}
\alias{geocodeDataFrame}
\title{Geocorde DataFrame}
\usage{
geocodeDataFrame(places, replace = FALSE)
}
\arguments{
\item{places}{a dataframe with at least one column containing the userLocation and google API}

\item{replace}{bolean value to define if existing lon/lat values should be replaced}
}
\value{
dataframe with coordinates (lat,long)
}
\description{
Converts a UserLocation to lat/lon coordinates
}
\examples{
#geocodeDataFrame(df)

}

