% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regexBuilder.R
\name{getMongoFields}
\alias{getMongoFields}
\title{getMongoFields}
\usage{
getMongoFields(fields)
}
\arguments{
\item{fields}{character vector of fields to be returned by a mongoDB query}
}
\value{
JSON format list of field to be returned
}
\description{
Creates JSON formated string of fields to be returned by a mongo find query,
so one does not have to write a JSON string "by hand".
}
\examples{
#getMongoFields(c("text","created_at"))

#getInterval(60*24)), getMongoFields(c("text","entitites.hashtags.text","created_at")))
}

