% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentification.R
\name{makeOAuth}
\alias{makeOAuth}
\title{Make OAuth}
\usage{
makeOAuth(key = Sys.getenv("twitter_key"),
  secret = Sys.getenv("twitter_secret"))
}
\arguments{
\item{key}{The key for the Twitter-API (character)}

\item{secret}{The scret for the Twitter-API (character)}
}
\value{
OAuth token
}
\description{
Creates oAuth token for Twitter-API Access and returns token
}
\examples{
#makeOAuth()

}

