% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regexBuilder.R
\name{operatorValuePair}
\alias{operatorValuePair}
\title{operatorValuePair}
\usage{
operatorValuePair(key, value)
}
\arguments{
\item{key}{of the key/value pair. Is always a mongo operator like $regex}

\item{value}{of the key/value pair. Can be the regex and/or the corresponding option.}
}
\value{
operator value pair
}
\description{
Creates operator value pair, called by buildRegexQuery.
}
\examples{
#operatorValuePair(c("$regex","$options"),c("koln","i"))

}

