% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongoQueries.R
\name{plotTweets}
\alias{plotTweets}
\title{plotTweets}
\usage{
plotTweets(data, tophashtags)
}
\arguments{
\item{data}{dataframe to plot created by createTwitterTimeSeries()}

\item{tophashtags}{number of hashtags to plot.}
}
\value{
plot of the given dataframe
}
\description{
Plots a dataframe in order to visualize the sum of hashtags in a
given time interval.
}
\examples{
#plots the top 10 hashtags
#plotTweets(data,10)
}

