% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongoQueries.R
\name{rankTweets}
\alias{rankTweets}
\title{rankTweets}
\usage{
rankTweets(data, interval)
}
\arguments{
\item{data}{dataframe time series to rank.}

\item{interval}{vector}
}
\value{
dataframe consisting of ranked hashtags
}
\description{
Ranks the hashtags according to their respective counts. Highest
count in a given interval is being assigned rank 1. Second highest
counts rank 2 and so on.
}
\examples{
#rankTweets(dataframe, interval)
}

