% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentification.R
\name{setAPICredentials}
\alias{setAPICredentials}
\title{Store Twitter-API Credentials}
\usage{
setAPICredentials(key, secret)
}
\arguments{
\item{key}{The key for the Twitter-API: Consumer Key (API Key)}

\item{secret}{The scret for the Twitter-API: Consumer Secret (API Secret)}
}
\value{
None
}
\description{
This function stores you Twitter-API key and secret to make it available for other functions and
for other instances of R.
}
\examples{
setAPICredentials(key = "Consumer Key", secret = "Consumer Secret")

}

