% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterstream.R
\name{startStreaming}
\alias{startStreaming}
\title{Start streaming}
\usage{
startStreaming(timeout = 60, query_url)
}
\arguments{
\item{timeout}{period ofg time after which the stream ist stoped
and restarted with a updated list of search parameters}

\item{query_url}{set a specific URL to query the API}
}
\value{
list of hashtags
}
\description{
Start the streaming process for the filterstream
in a seperate R process
}
\examples{
#startStreaming(320)
#startStreaming(60,"https://stream.twitter.com/1.1/statuses/filter.json?track=Syrieneinsatz")

}

