% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterstream.R
\name{stopStreaming}
\alias{stopStreaming}
\title{Stop streaming}
\usage{
stopStreaming()
}
\value{
list of hashtags
}
\description{
Stop the streaming process. Streaming is stoped within the timeframe given
as timeout in \link{startStreaming}
Stop stream will not terminate the streaming process imediately but stop instead
}
\examples{
#stopStreaming()

}

