% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateOperation.R
\name{topHashtagsInInterval}
\alias{topHashtagsInInterval}
\title{Top Hashtags in intervall}
\usage{
topHashtagsInInterval(mongo, interval, amount = 10)
}
\arguments{
\item{mongo}{a mongoDB connection to use (default: hometimeline)}

\item{interval}{a time interval which can be generated by \link{getInterval} needs to be in unix timestamp format}

\item{amount}{limits the number of hashtags returned}
}
\value{
list of hashtags
}
\description{
Retrives the most used hashtags in a given time interval
}
\examples{
#topHashtagsInInterval(hometimeline, interval = getInterval(60*24), 10)
#or alternatively specify interval beforehand
#timeframe <- c("2016-01-04 18:37:43 CET","2016-01-05 18:37:43 CET")
#hashtags <- topHashtagsInInterval(hometimeline, interval =
#getInterval(inputinterval = timeframe), 10)

}

