% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hometimeline.R
\name{updateHometimeline}
\alias{updateHometimeline}
\title{Update Home Timeline}
\usage{
updateHometimeline(connection = "hometimeline", oAuth = makeOAuth())
}
\arguments{
\item{connection}{a mongoconnection to use (default: hometimeline)}

\item{oAuth}{an oAuth token to use, by default a token is retrived by using \link{makeOAuth}}
}
\value{
None
}
\description{
Function to retrive and update the home timeline, the home timeline consist of up to 800 tweets
posted by the accounts the user follows. Set Sys.setlocale("LC_TIME", "English") in order
to be able to correctly process the twitter timestamp. See ?as.POSIXct for help.
}
\examples{
#updateHometimeline()

}

