% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regexBuilder.R
\name{useConjunction}
\alias{useConjunction}
\title{useConjunction}
\usage{
useConjunction(firstpart, secondpart, conjunction)
}
\arguments{
\item{firstpart}{first part of the query}

\item{secondpart}{second part of the query}

\item{conjunction}{which can only be "$and" or "$or"}
}
\value{
linked mongo query parts.
}
\description{
Connects query parts with either $and or $or.
}
\examples{
useConjunction(c("$regex","$options"),c("syrien","i"),"$and")
}

