% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongoQueries.R
\name{plotTweets}
\alias{plotTweets}
\title{plotTweets}
\usage{
plotTweets(data, tophashtags, interval)
}
\arguments{
\item{data}{dataframe to plot created by createTwitterTimeSeries()}

\item{tophashtags}{number of hashtags to plot.}

\item{interval}{specified time interval. Relevant for the x-axis of the plot.}
}
\value{
plot of the given dataframe
}
\description{
Plots a dataframe in order to visualize the sum of hashtags in a
given time interval.
}
\examples{
#plots the top 10 hashtags
#plotTweets(data,10)
#interval <- createIntervalVector(c("2016-01-15 20:00:00", "2016-01-16 00:10:00"),9)
#interval
#plotTweets(data,10, interval)
}

