% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterstream.R
\name{startStreaming}
\alias{startStreaming}
\title{Start streaming}
\usage{
startStreaming(timeout = 60, query_url)
}
\arguments{
\item{timeout}{period of time after which the stream ist stopped
and restarted with a updated list of search parameters}

\item{query_url}{set a specific URL to query the API}
}
\value{
list of hashtags
}
\description{
Start the streaming process for the filterstream
in a seperate R process. The timeout also defines the maximum waiting time
between executing \link{stopStreaming} and the end of the streaming process
If no query_url is supplied then the topHashtags from the users hometimeline
are used.
}
\examples{
#startStreaming(320)
#startStreaming(60,"https://stream.twitter.com/1.1/statuses/filter.json?track=Trump")

}

